requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.13
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0

pypdf~=5.1.0
